+++ src/fixtures.constants.ts
@@ -1,0 +1,1 @@
+export const isExtension = process.env.VITEST_VSCODE === 'true';