+++ src/reduceFunctions.types.ts
@@ -1,5 +1,5 @@
-import type { LengthOf, TupleOf } from '@bemedev/types';
-import type { Checker_F } from './types';
+import type { TupleOf } from '@bemedev/core/lib/globals/types.js';
+import type { Checker_F, LengthOf } from './types';
 
 export type ReduceFunctions_F = <const T extends Checker_F[]>(
   ...functions: T